refactor: 데이터 타입 수정
- 엔티티의 필드 타입에 맞게 Id 값의 타입은 Reference Type 사용
- 시간 또는 카운트 숫자의 데이터 경우 성능을 고려해서 Primitive Type 사용

+++ src/test/java/coffeemeet/server/auth/domain/JwtTokenProviderTest.java
@@ -15,7 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class JwtTokenProviderTest {
 
   private static final String SUBJECT = "123";
-  private static final int ACCESS_TOKEN_EXPIRE_TIME = 3600000;
+  private static final long ACCESS_TOKEN_EXPIRE_TIME = 3600000;
 
   private final String secretKey = Base64.getEncoder()
       .encodeToString(Keys.secretKeyFor(SignatureAlgorithm.HS512).getEncoded());
